+++ R/ewce_para.R
@@ -73,7 +73,7 @@ ewce_para <- function(ctd,
     i <- which(list_names==p)
     genes <- gene_lists[[p]]
     message_parallel("Analysing: ",shQuote(p),
-                     " (",i,"/",length(list_names),"):",
+                     " (",i,"/",length(list_names),"): ",
                      formatC(length(genes),big.mark = ",")," genes")
     tryCatch({
       results <- EWCE::bootstrap_enrichment_test(