+++ R/get_unfinished_list_names.R
@@ -14,22 +14,20 @@
 #' gene_data <- HPOExplorer::load_phenotype_to_genes()
 #' list_names <- unique(gene_data$Phenotype)[seq_len(3)]
 #' save_dir_tmp <- file.path(tempdir(),"results")
+#' ctd <- load_example_ctd()
 #' res_files <- ewce_para(ctd = ctd,
 #'                        gene_data = gene_data,
 #'                        list_names = list_names,
 #'                        reps = 10,
 #'                        save_dir_tmp = save_dir_tmp)
-#' unfinished <- get_unfinished_list_names(list_names = list_names,
+#' unfinished <- get_unfinished_list_names(list_names = gene_data$Phenotype,
 #'                                         save_dir_tmp = save_dir_tmp)
 get_unfinished_list_names <- function (list_names,
                                        save_dir_tmp) {
+  list_names <- unique(list_names)
   if(is.null(save_dir_tmp)) return(list_names)
-  lapply(list_names, function(l){
-    if (is_not_analysed(list_name = l,
-                        save_dir_tmp = save_dir_tmp)) {
-      return(l)
-    } else {
-      return(NULL)
-    }
-  }) |> unlist()
+  unfinished <- list_names[is_not_analysed(list_names = list_names,
+                                            save_dir_tmp = save_dir_tmp)
+                          ]
+  return(unfinished)
 }