+++ R/is_not_analysed.R
@@ -5,14 +5,13 @@
 #' The \link[MultiEWCE]{ewce_para} function outputs the results
 #' for each gene list in to the results directory (\code{save_dir_tmp})
 #'  with name format \emph{list_name.rds}.
-#' @param list_name The name of a gene list (e.g. "Phenotypic abnormality").
 #' @inheritParams ewce_para
-#' @returns TRUE or FALSE \code{bool}.
+#' @returns Vector of TRUE or FALSE.
 
 #' @keywords internal
-is_not_analysed <- function(list_name,
+is_not_analysed <- function(list_names,
                             save_dir_tmp) {
   file_path <- make_save_path(save_dir = save_dir_tmp,
-                              list_name = list_name)
+                              list_name = list_names)
   return(!file.exists(file_path))
 }